app/backend/requirements.txt
fastapi==0.114.0
uvicorn==0.30.6
google-cloud-texttospeech==2.16.4
google-auth==2.33.0
google-cloud-secret-manager==2.20.2
